+++ sdk/resourcemanager/datafactory/armdatafactory/models.go
@@ -3672,7 +3672,7 @@ type AzureFunctionActivityTypeProperties struct {
 
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
-	// string (or Expression with resultType string).
+	// dictionary (or Expression with resultType dictionary).
 	Headers any
 }
 
@@ -3714,7 +3714,8 @@ func (a *AzureFunctionLinkedService) GetLinkedService() *LinkedService {
 
 // AzureFunctionLinkedServiceTypeProperties - Azure Function linked service properties.
 type AzureFunctionLinkedServiceTypeProperties struct {
-	// REQUIRED; The endpoint of the Azure Function App. URL will be in the format https://.azurewebsites.net.
+	// REQUIRED; The endpoint of the Azure Function App. URL will be in the format https://.azurewebsites.net. Type: string (or
+	// Expression with resultType string).
 	FunctionAppURL any
 
 	// Type of authentication (Required to specify MSI) used to connect to AzureFunction. Type: string (or Expression with resultType
@@ -3731,7 +3732,7 @@ type AzureFunctionLinkedServiceTypeProperties struct {
 	// Function or Host key for Azure Function App.
 	FunctionKey SecretBaseClassification
 
-	// Allowed token audiences for azure function.
+	// Allowed token audiences for azure function. Type: string (or Expression with resultType string).
 	ResourceID any
 }
 
@@ -3964,8 +3965,8 @@ type AzureMLExecutePipelineActivityTypeProperties struct {
 	ContinueOnStepFailure any
 
 	// Dictionary used for changing data path assignments without retraining. Values will be passed in the dataPathAssignments
-	// property of the published pipeline execution request. Type: object with key
-	// value pairs (or Expression with resultType object).
+	// property of the published pipeline execution request. Type: object (or
+	// Expression with resultType object).
 	DataPathAssignments any
 
 	// Run history experiment name of the pipeline run. This information will be passed in the ExperimentName property of the
@@ -5258,7 +5259,7 @@ type AzureSQLSource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -6546,10 +6547,10 @@ type CmdkeySetupTypeProperties struct {
 	// REQUIRED; The password of data source access.
 	Password SecretBaseClassification
 
-	// REQUIRED; The server name of data source access.
+	// REQUIRED; The server name of data source access. Type: string.
 	TargetName any
 
-	// REQUIRED; The user name of data source access.
+	// REQUIRED; The user name of data source access. Type: string.
 	UserName any
 }
 
@@ -13713,8 +13714,7 @@ type HDInsightOnDemandLinkedServiceTypeProperties struct {
 	// REQUIRED; The resource group where the cluster belongs. Type: string (or Expression with resultType string).
 	ClusterResourceGroup any
 
-	// REQUIRED; Number of worker/data nodes in the cluster. Suggestion value: 4. Type: string (or Expression with resultType
-	// string).
+	// REQUIRED; Number of worker/data nodes in the cluster. Suggestion value: 4. Type: int (or Expression with resultType int).
 	ClusterSize any
 
 	// REQUIRED; The customer’s subscription to host the cluster. Type: string (or Expression with resultType string).
@@ -14226,8 +14226,8 @@ type HTTPLinkedServiceTypeProperties struct {
 	// string).
 	URL any
 
-	// The additional HTTP headers in the request to RESTful API used for authorization. Type: object (or Expression with resultType
-	// object).
+	// The additional HTTP headers in the request to RESTful API used for authorization. Type: key value pairs (value should be
+	// string type).
 	AuthHeaders any
 
 	// The authentication type to be used to connect to the HTTP server.
@@ -17498,12 +17498,28 @@ type MariaDBLinkedServiceTypeProperties struct {
 	// An ODBC connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
 	ConnectionString any
 
+	// Database name for connection. Type: string.
+	Database any
+
+	// The version of the MariaDB driver. Type: string. V1 or empty for legacy driver, V2 for new driver. V1 can support connection
+	// string and property bag, V2 can only support connection string.
+	DriverVersion any
+
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
 	// The Azure key vault secret reference of password in connection string.
-	Pwd *AzureKeyVaultSecretReference
+	Password *AzureKeyVaultSecretReference
+
+	// The port for the connection. Type: integer.
+	Port any
+
+	// Server name for connection. Type: string.
+	Server any
+
+	// Username for authentication. Type: string.
+	Username any
 }
 
 // MariaDBSource - A copy activity MariaDB server source.
@@ -18654,8 +18670,15 @@ func (m *MySQLLinkedService) GetLinkedService() *LinkedService {
 
 // MySQLLinkedServiceTypeProperties - MySQL linked service properties.
 type MySQLLinkedServiceTypeProperties struct {
-	// REQUIRED; The connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
+	// The connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
 	ConnectionString any
+
+	// Database name for connection. Type: string.
+	Database any
+
+	// The version of the MySQL driver. Type: string. V1 or empty for legacy driver, V2 for new driver. V1 can support connection
+	// string and property bag, V2 can only support connection string.
+	DriverVersion any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
@@ -18663,6 +18686,21 @@ type MySQLLinkedServiceTypeProperties struct {
 
 	// The Azure key vault secret reference of password in connection string.
 	Password *AzureKeyVaultSecretReference
+
+	// The port for the connection. Type: integer.
+	Port any
+
+	// SSL mode for connection. Type: integer. 0: disable, 1: prefer, 2: require, 3: verify-ca, 4: verify-full.
+	SSLMode any
+
+	// Server name for connection. Type: string.
+	Server any
+
+	// Use system trust store for connection. Type: integer. 0: enable, 1: disable.
+	UseSystemTrustStore any
+
+	// Username for authentication. Type: string.
+	Username any
 }
 
 // MySQLSource - A copy activity source for MySQL databases.
@@ -19021,8 +19059,8 @@ type ODataLinkedServiceTypeProperties struct {
 	// Specify the credential type (key or cert) is used for service principal.
 	AADServicePrincipalCredentialType *ODataAADServicePrincipalCredentialType
 
-	// The additional HTTP headers in the request to RESTful API used for authorization. Type: object (or Expression with resultType
-	// object).
+	// The additional HTTP headers in the request to RESTful API used for authorization. Type: key value pairs (value should be
+	// string type).
 	AuthHeaders any
 
 	// Type of authentication used to connect to the OData service.
@@ -22552,7 +22590,7 @@ type RestSink struct {
 	// REQUIRED; Copy sink type.
 	Type *string
 
-	// The additional HTTP headers in the request to the RESTful API. Type: string (or Expression with resultType string).
+	// The additional HTTP headers in the request to the RESTful API. Type: key value pairs (value should be string type).
 	AdditionalHeaders any
 
 	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
@@ -22562,7 +22600,7 @@ type RestSink struct {
 	DisableMetricsCollection any
 
 	// Http Compression Type to Send data in compressed format with Optimal Compression Level, Default is None. And The Only Supported
-	// option is Gzip.
+	// option is Gzip. Type: string (or Expression with resultType string).
 	HTTPCompressionType any
 
 	// The timeout (TimeSpan) to get an HTTP response. It is the timeout to get a response, not the timeout to read response data.
@@ -22611,8 +22649,7 @@ type RestSource struct {
 	// REQUIRED; Copy source type.
 	Type *string
 
-	// Specifies the additional columns to be added to source data. Type: array of objects(AdditionalColumns) (or Expression with
-	// resultType array of objects).
+	// Specifies the additional columns to be added to source data. Type: key value pairs (value should be string type).
 	AdditionalColumns any
 
 	// The additional HTTP headers in the request to the RESTful API. Type: string (or Expression with resultType string).
@@ -22825,7 +22862,7 @@ type SQLDWSource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -22938,7 +22975,7 @@ type SQLMISink struct {
 	// Write batch timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
 	WriteBatchTimeout any
 
-	// White behavior when copying data into azure SQL MI. Type: SqlWriteBehaviorEnum (or Expression with resultType SqlWriteBehaviorEnum)
+	// White behavior when copying data into azure SQL MI. Type: string (or Expression with resultType string)
 	WriteBehavior any
 }
 
@@ -22980,7 +23017,7 @@ type SQLMISource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -23159,7 +23196,7 @@ type SQLServerSink struct {
 	// Write batch timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
 	WriteBatchTimeout any
 
-	// Write behavior when copying data into sql server. Type: SqlWriteBehaviorEnum (or Expression with resultType SqlWriteBehaviorEnum)
+	// Write behavior when copying data into sql server. Type: string (or Expression with resultType string).
 	WriteBehavior any
 }
 
@@ -23201,7 +23238,7 @@ type SQLServerSource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -23444,7 +23481,7 @@ type SQLSink struct {
 	// Write batch timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
 	WriteBatchTimeout any
 
-	// Write behavior when copying data into sql. Type: SqlWriteBehaviorEnum (or Expression with resultType SqlWriteBehaviorEnum)
+	// Write behavior when copying data into sql. Type: string (or Expression with resultType string).
 	WriteBehavior any
 }
 
@@ -23486,7 +23523,7 @@ type SQLSource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -29765,13 +29802,23 @@ type WebActivityTypeProperties struct {
 	// When set to true, Certificate validation will be disabled.
 	DisableCertValidation *bool
 
+	// Timeout for the HTTP request to get a response. Format is in TimeSpan (hh:mm:ss). This value is the timeout to get a response,
+	// not the activity timeout. The default value is 00:01:00 (1 minute). The
+	// range is from 1 to 10 minutes
+	HTTPRequestTimeout any
+
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
-	// string (or Expression with resultType string).
+	// dictionary (or Expression with resultType dictionary).
 	Headers any
 
 	// List of linked services passed to web endpoint.
 	LinkedServices []*LinkedServiceReference
+
+	// Option to disable invoking HTTP GET on location given in response header of a HTTP 202 Response. If set true, it stops
+	// invoking HTTP GET on http location given in response header. If set false then
+	// continues to invoke HTTP GET call on location given in http response headers.
+	TurnOffAsync *bool
 }
 
 // WebAnonymousAuthentication - A WebLinkedService that uses anonymous authentication to communicate with an HTTP endpoint.
@@ -29921,7 +29968,7 @@ type WebHookActivityTypeProperties struct {
 
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
-	// string (or Expression with resultType string).
+	// dictionary (or Expression with resultType dictionary).
 	Headers any
 
 	// When set to true, statusCode, output and error in callback request body will be consumed by activity. The activity can